refactor: consolidate API by moving benchmark() to Evaluation.jl
- Move benchmark() function from Interface.jl to Evaluation.jl  
- Delete src/Interface.jl (no longer needed)
- Relocate tests:
  - Move Benchmark testset to test/Evaluation.jl
  - Move attack dispatch tests to test/attacks/Attack.jl
  - Delete test/Interface.jl
- Update module exports in src/AdversarialAttacks.jl
- Merge docs/src/interface.md into docs/src/attack_interface.md

Fixes test failures:
- Rename DummyAttack structs to avoid redefinition conflicts
- Scope DecisionTree imports locally to prevent fit! conflict with MLJ

All tests pass (95 tests). Documentation builds successfully.
