+++ docs/src/evaluation.md
@@ -16,4 +16,5 @@ AdversarialAttacks.calculate_metrics(n_test, num_clean_correct, num_adv_correct,
 AdversarialAttacks.compute_norm(sample_data, adv_data, p::Real)
 AdversarialAttacks.evaluate_robustness(model,attack,test_data;num_samples::Int=100)
 AdversarialAttacks.evaluation_curve(model, atk_type::Type{<:AbstractAttack}, epsilons::Vector{Float64}, test_data; num_samples::Int = 100)
+AdversarialAttacks.benchmark
 ```
