+++ test/runtests.jl
@@ -5,4 +5,3 @@ include("attacks/Attack.jl")
 include("attacks/FGSM.jl")
 include("attacks/BasicRandomSearch.jl")
 include("Evaluation.jl")
-include("Interface.jl")
